Move FP16 tests to a dedicated file
